M3-198 Chore : ApplePublicKeyGenerator 공통 주석 추가
